+++ package.json
@@ -1,7 +1,7 @@
 {
 	"name": "elysia",
 	"description": "Ergonomic Framework for Human",
-	"version": "1.3.0-exp.3",
+	"version": "1.3.0-exp.4",
 	"author": {
 		"name": "saltyAom",
 		"url": "https://github.com/SaltyAom",
@@ -53,6 +53,11 @@
 			"import": "./dist/error.mjs",
 			"require": "./dist/cjs/error.js"
 		},
+		"./schema": {
+			"types": "./dist/schema.d.ts",
+			"import": "./dist/schema.mjs",
+			"require": "./dist/cjs/schema.js"
+		},
 		"./sucrose": {
 			"types": "./dist/sucrose.d.ts",
 			"import": "./dist/sucrose.mjs",
@@ -102,6 +107,11 @@
 			"types": "./dist/adapter/bun/handler.d.ts",
 			"import": "./dist/adapter/bun/handler.mjs",
 			"require": "./dist/cjs/adapter/bun/handler.js"
+		},
+		"./adapter/bun/compose": {
+			"types": "./dist/adapter/bun/compose.d.ts",
+			"import": "./dist/adapter/bun/compose.mjs",
+			"require": "./dist/cjs/adapter/bun/compose.js"
 		},
 		"./adapter/web-standard": {
 			"types": "./dist/adapter/web-standard/index.d.ts",
